refactor(nmp-common): split plugin-safe SDK factory + helpers into nemo_platform_plugin
Phase B/C of the nmp-common slim plan (follow-up to ASTD-166).

Canonical homes for plugin-safe helpers now live in nemo_platform_plugin:
- jobs/image.py     (get_qualified_image, image_builder)
- http_clients.py   (shared sync/async HTTP clients)
- otel_headers.py   (MARK_INTERNAL_REQUEST_HEADERS, propagation context)
- auth/             (Principal, AuthContext, principal_from_env,
                     auth_client_context, get_principal_auth_headers,
                     build_service_principal_headers, exceptions)
- sdk_factory.py    (get_platform_sdk, get_async_platform_sdk, get_task_sdk,
                     get_request_scoped_sdk, get_sdk_on_behalf_of,
                     get_entity_parts)

nmp.common modules are thin shims that re-export from these locations so
service-side callers keep working. Plugin call sites in nemo-{anonymizer,
data-designer,guardrails,agents,example-plugin} now import directly from
nemo_platform_plugin.

Also tightens NemoJob.to_spec / NemoJob.compile base signatures: async_sdk
from object to AsyncNeMoPlatform. Resolves Liskov override violations
every concrete subclass was already hitting.

nmp_testing.client and test_sdk_factory updated to mutate the canonical
_test_http_client and patch the canonical module path.

No install-footprint change yet - plugins still import other nmp.common
helpers (entities.*, config.*, service.headers.*). Phase D removes the dep
once those land.
